Add new flavor python3-ds-cuda-preview (#40)
This commit adds a new flavor which enables the usage of CUDA CPUs in the Container. This is currently specifically meant for Tensorflow, but will be extended to other libraries.
The flavor installs the CUDA Driver Version 410.104 without the Kernel module and the CUDA SDK Version 10 into the Container. These versions are specifically chosen for Tensorflow for GPUs in Version 1.13.1, because Tensorflow installed with pip depends on specific version of libcuda.so. To use Tensorflow with GPUs the Host system needs the complete CUDA Driver with the same version installed.

NOTE: Currently, Travis CI is not able to upload the pre-packaged python3-ds-cuda-preview container to Github, because it fails with the following Error: travis-ci/travis-ci#8553

Further, minor changes:

- bugfixes for the exaslct starter script on some installations of python3
- bugfix of the travis.yaml which now pushes all images to the build caches during deployment
- added PyYAML to all python3-ds-* flavors