Bugfixes for exaslct starter script and fixing tensorflow version to 1.13.1 (#44)
This commit fixes bugs in the exaslct starter script for the case that exaslct doesn't get called from the repository root. Further, it fixes the tensorflow version to 1.13.1, because tensorflow 1.14.0 and probably tensorflow 2.0 raise an error during import in the UDF. See, more details https://github.com/exasol/script-languages/issues/43.