Handle multiple lines at once in LogHandlers (#136)
Docker on MacOSX seems to return multiple lines at once when streaming API call logs. This patch extends the API of all LogHandlers with the function handle_log_lines to accept multiple lines. The AbstractLogHandler does the decoding and splitting of the lines and feeds them into the method handle_log_line.
The DockerRegistryImageChecker used before its own method to parse the log, it now uses a specific implementation of the LogHandler which simplifies this step.